+++ coderadar-core/src/main/java/io/reflectoring/coderadar/vcs/port/driven/DeleteLocalRepositoryPort.java
@@ -1,7 +1,7 @@
-package io.reflectoring.coderadar.vcs.port.driver;
+package io.reflectoring.coderadar.vcs.port.driven;
 
 import java.io.IOException;
 
-public interface DeleteRepositoryUseCase {
+public interface DeleteLocalRepositoryPort {
   void deleteRepository(String repositoryPath) throws IOException;
 }
