--- coderadar-core/src/main/java/io/reflectoring/coderadar/vcs/port/driven/DeleteRepositoryPort.java
@@ -1,7 +1,0 @@
-package io.reflectoring.coderadar.vcs.port.driven;
-
-import java.io.IOException;
-
-public interface DeleteRepositoryPort {
-  void deleteRepository(String repositoryPath) throws IOException;
-}
