+++ coderadar-rest/src/main/java/io/reflectoring/coderadar/rest/ExceptionHandlerController.java
@@ -14,22 +14,22 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 public class ExceptionHandlerController {
 
     @ExceptionHandler(ProjectIsBeingProcessedException.class)
-    public ResponseEntity projectProcessingException(ProjectIsBeingProcessedException e) {
+    public ResponseEntity<ErrorMessageResponse> projectProcessingException(ProjectIsBeingProcessedException e) {
         return new ResponseEntity<>(new ErrorMessageResponse(e.getMessage()), HttpStatus.UNPROCESSABLE_ENTITY);
     }
 
     @ExceptionHandler(EntityNotFoundException.class)
-    public ResponseEntity entityNotFoundException(EntityNotFoundException e) {
+    public ResponseEntity<ErrorMessageResponse> entityNotFoundException(EntityNotFoundException e) {
         return new ResponseEntity<>(new ErrorMessageResponse(e.getMessage()), HttpStatus.NOT_FOUND);
     }
 
     @ExceptionHandler(EntityAlreadyExistsException.class)
-    public ResponseEntity entityAlreadyExistsException(EntityAlreadyExistsException e) {
+    public ResponseEntity<ErrorMessageResponse> entityAlreadyExistsException(EntityAlreadyExistsException e) {
         return new ResponseEntity<>(new ErrorMessageResponse(e.getMessage()), HttpStatus.CONFLICT);
     }
 
     @ExceptionHandler(MisconfigurationException.class)
-    public ResponseEntity misconfigurationException(MisconfigurationException e) {
+    public ResponseEntity<ErrorMessageResponse> misconfigurationException(MisconfigurationException e) {
         return new ResponseEntity<>(new ErrorMessageResponse(e.getMessage()), HttpStatus.UNPROCESSABLE_ENTITY);
     }
 }
